+++ src/candidate_openmp_v2/candidate_openmp_v2_text_search.cpp
@@ -58,7 +58,7 @@ find_candidate_openmp_v2(const std::string &text,
 
 #pragma omp parallel for default(none)                                         \
     shared(queries, mask_words, mask, text, indices)
-    for (long long i = 0; i < queries.size(); ++i) {
+    for (long i = 0; i < queries.size(); ++i) {
         const std::string &query = queries[i];
 
         for (unsigned long word = 0; word < mask_words; ++word) {
